build test for: Code, Layout, App, MissionItem
