+++ src/components/MissionItem.js
@@ -1,23 +1,11 @@
 import Badge from 'react-bootstrap/Badge';
 import Button from 'react-bootstrap/Button';
-import { useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
-import { joinMission, leaveMission } from '../redux/missions/missionsSlice';
 
 const MissionItem = (props) => {
-  const dispatch = useDispatch();
   const {
-    id, name, description, reserved,
+    id, name, description, reserved, handleClick,
   } = props;
-
-  const handleClick = (e, { id, reserved }) => {
-    e.preventDefault();
-    if (reserved) {
-      dispatch(leaveMission({ id }));
-    } else {
-      dispatch(joinMission({ id }));
-    }
-  };
 
   return (
     <tr>
@@ -29,6 +17,7 @@ const MissionItem = (props) => {
       </td>
       <td className="align-middle" style={{ width: '11rem', textAlign: 'center' }}>
         <Button
+          data-testid="mission-item-button"
           variant={reserved ? 'outline-danger' : 'outline-dark'}
           type="button"
           onClick={(e) => {
@@ -47,6 +36,7 @@ MissionItem.propTypes = {
   name: PropTypes.string.isRequired,
   description: PropTypes.string.isRequired,
   reserved: PropTypes.bool.isRequired,
+  handleClick: PropTypes.func.isRequired,
 };
 
 export default MissionItem;
